Fix tests in billing mappers/services
