+++ src/lib/connectors/repos/BillingBatchRepository.js
@@ -1,4 +1,5 @@
 const { BillingBatch } = require('../bookshelf');
+const { paginatedEnvelope } = require('./lib/envelope');
 
 const findOne = async (id) => {
   const model = await new BillingBatch({ billing_batch_id: id })
@@ -12,4 +13,19 @@ const findOne = async (id) => {
 }
 ;
 
+const findPage = async (page, pageSize) => {
+  const result = await new BillingBatch()
+    .orderBy('date_created', 'ASC')
+    .fetchPage({
+      page,
+      pageSize,
+      withRelated: [
+        'region'
+      ]
+    });
+  console.log(result);
+  return paginatedEnvelope(result);
+};
+
 exports.findOne = findOne;
+exports.findPage = findPage;
