+++ src/lib/connectors/repos/lib/envelope.js
@@ -1,0 +1,14 @@
+const { pick } = require('lodash');
+
+const mapBookshelfPagination = pagination => ({
+  ...pick(pagination, ['page', 'pageCount']),
+  perPage: pagination.pageSize,
+  totalRows: pagination.rowCount
+});
+
+const paginatedEnvelope = bookshelfResult => ({
+  data: bookshelfResult.toJSON(),
+  pagination: mapBookshelfPagination(bookshelfResult.pagination)
+});
+
+exports.paginatedEnvelope = paginatedEnvelope;
