+++ src/modules/billing/mappers/index.js
@@ -1,8 +1,8 @@
 exports.address = require('./address');
 exports.abstractionPeriod = require('./abstraction-period');
 exports.batch = require('./batch');
-exports.agreements = require('./agreements');
-exports.chargeElements = require('./charge-elements');
+exports.agreement = require('./agreement');
+exports.chargeElement = require('./charge-element');
 exports.company = require('./company');
 exports.contact = require('./contact');
 exports.invoice = require('./invoice');
@@ -10,4 +10,4 @@ exports.invoiceAccount = require('./invoice-account');
 exports.invoiceLicence = require('./invoice-licence');
 exports.licence = require('./licence');
 exports.region = require('./region');
-exports.transactions = require('./transactions');
+exports.transaction = require('./transaction');
