+++ src/modules/billing/services/batch-service.js
@@ -1,4 +1,3 @@
-const repos = require('../../../lib/connectors/repository');
 const newRepos = require('../../../lib/connectors/repos');
 const mappers = require('../mappers');
 
@@ -13,20 +12,10 @@ const getBatchById = async id => {
 };
 
 const getBatches = async (page = 1, perPage = Number.MAX_SAFE_INTEGER) => {
-  const pagination = { page, perPage };
-  const { rows } = await repos.billingBatches.find(null, { date_created: -1 }, pagination);
-  const batches = rows.map(mappers.batch.dbToModel);
-
-  const { rows: [totalRowCount] } = await repos.billingBatches.findRowCount();
-  const totalRows = totalRowCount.totalrowcount;
+  const result = await newRepos.billingBatches.findPage(page, perPage);
   return {
-    data: batches,
-    pagination: {
-      page,
-      perPage,
-      totalRows,
-      pageCount: Math.ceil(totalRows / perPage)
-    }
+    data: result.data.map(mappers.batch.dbToModel),
+    pagination: result.pagination
   };
 };
