+++ test/modules/billing/mappers/charge-element.js
@@ -1,15 +1,12 @@
 const {
   experiment,
   test,
-  beforeEach,
-  afterEach
+  beforeEach
 } = exports.lab = require('@hapi/lab').script();
 const { expect } = require('@hapi/code');
-const sandbox = require('sinon').createSandbox();
 
 const ChargeElement = require('../../../../src/lib/models/charge-element');
-const chargeElementsService = require('../../../../src/modules/billing/services/charge-elements-service');
-const repos = require('../../../../src/lib/connectors/repository');
+const chargeElementsMapper = require('../../../../src/modules/billing/mappers/charge-element');
 
 const data = {
   chargeElement: {
@@ -26,21 +23,38 @@ const data = {
   }
 };
 
-experiment('modules/billing/services/charge-elements-service', () => {
+experiment('modules/billing/mappers/charge-element', () => {
   let result;
 
-  beforeEach(async () => {
-    sandbox.stub(repos.chargeElements, 'findOneById');
-  });
-
-  afterEach(async () => {
-    sandbox.restore();
-  });
-
-  experiment('.getById', () => {
+  experiment('.chargeToModel', () => {
     beforeEach(async () => {
-      repos.chargeElements.findOneById.resolves(data.dbRow);
-      result = await chargeElementsService.getById(data.dbRow);
+      result = chargeElementsMapper.chargeToModel(data.chargeElement);
+    });
+
+    test('returns an instance of ChargeElement', async () => {
+      expect(result instanceof ChargeElement).to.be.true();
+    });
+
+    test('sets the .id property', async () => {
+      expect(result.id).to.equal(data.chargeElement.chargeElementId);
+    });
+
+    test('sets the .source property', async () => {
+      expect(result.source).to.equal(data.chargeElement.source);
+    });
+
+    test('sets the .season property', async () => {
+      expect(result.season).to.equal(data.chargeElement.season);
+    });
+
+    test('sets the .loss property', async () => {
+      expect(result.loss).to.equal(data.chargeElement.loss);
+    });
+  });
+
+  experiment('.dbToModel', () => {
+    beforeEach(async () => {
+      result = chargeElementsMapper.dbToModel(data.dbRow);
     });
 
     test('returns an instance of ChargeElement', async () => {
